Fix handling of negative bytes in input
Negative bytes were converted to UIntas is, without consideration for
their sign (UInt is higher precision and can fit both). Converting to
UByte first ensures sign is handled (i.e., masked with 0xff).

This commit also increases test coverage of edge cases, by testing with
all bytes in random order, similar to Apache's Commons Codec test suite.

Closes #3

+++ src/commonMain/kotlin/MurmurHash3.kt
@@ -295,9 +295,9 @@ public class MurmurHash3(private val seed: UInt = 0u) {
         return h
     }
 
-    private fun ByteArray.getUInt(index: Int) = get(index).toUInt()
+    private fun ByteArray.getUInt(index: Int) = get(index).toUByte().toUInt()
 
-    private fun ByteArray.getULong(index: Int) = get(index).toULong()
+    private fun ByteArray.getULong(index: Int) = get(index).toUByte().toULong()
 
     private companion object {
         private const val C1_32: UInt = 0xcc9e2d51u
